refactor: split layout into components and update imports
