+++ src/components/features/eBook/layout/index.ts
@@ -1,0 +1,4 @@
+export { default as Section } from './Section';
+export { default as Figure } from './Figure';
+export { default as MaxWidthContainer } from './MaxWidthContainer';
+export { default as InlineDivs } from './InlineDivs';
